pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-agent-cli"
version = "0.2.0"
description = "CLI tool for interacting with Ollama as a cybersecurity agent with tool calling"
requires-python = ">=3.10"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "ollama>=0.1.35",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=0.42b0",
    "opentelemetry-instrumentation>=0.42b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ollama-agent = "ollama_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages]
find = {where = ["src"]}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
